Final Update

+++ tiffin/src/app/merchant-home/merchant-home.component.ts
@@ -1,10 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { Router } from '@angular/router';
-import { FoodService } from 'src/services/food/food.service';
 import { MerchantService } from 'src/services/merchant/merchant.service';
 import { TokenStorageService } from 'src/services/token-storage/token-storage.service';
-import { FoodItems } from '../shared/models/foodItem';
 
 @Component({
   selector: 'app-merchant-home',
@@ -13,18 +11,14 @@ import { FoodItems } from '../shared/models/foodItem';
 })
 export class MerchantHomeComponent implements OnInit {
 
-  foods: FoodItems[] = [];
   foodItems !: any;
 
   constructor(private tokenStorage: TokenStorageService, 
-    private merchantService: MerchantService, private foodItem: FoodService, 
+    private merchantService: MerchantService, 
     private router: Router, private sanitizer: DomSanitizer) {
      }
 
   ngOnInit(): void {
-    // console.log(window.sessionStorage);
-    // console.log(this.tokenStorage.getMerchantFoodItems());
-    // console.log(this.tokenStorage.getMerchantFoodItems().length);
     const items =  this.tokenStorage.getMerchantFoodItems();
     console.log(items);
     if(items !== null){
@@ -38,8 +32,6 @@ export class MerchantHomeComponent implements OnInit {
     console.log(items);
     this.foodItems = items;
     console.log(this.tokenStorage.getMerchantInfo())
-    // this.foods = this.foodItem.getAll();
-    // this.foodItems = this.tokenStorage.getMerchantFoodItems();
   }
 
   addFoodItem(){
@@ -48,29 +40,33 @@ export class MerchantHomeComponent implements OnInit {
 
   deleteFoodItem(itemId: number){
     console.log(itemId);
-    this.merchantService
+    const result = confirm('Are you sure you want to delete this item?');
+    if(!!result) {
+      this.merchantService
       .deleteFoodItem(itemId)
       .subscribe((res: any) => {
         console.log(res);
         this.tokenStorage.saveMerchantFoodItems(res);
         window.location.reload();
       });
+    }
   }
 
   editFoodItem(foodItemDetails: any) {
     this.merchantService.setEditFoodItemDetails(foodItemDetails);
     this.router.navigate(['edit-fooditem']);
-    // this.merchantService
-    //   .editFoodItem(itemId)
-    //   .subscribe((res: any) => {
-    //     console.log(res);
-    //     this.tokenStorage.saveMerchantFoodItems(res);
-    //     window.location.reload();
-    //   });
   }
 
   getSanitizeUrl(url: string){
     return this.sanitizer.bypassSecurityTrustResourceUrl(url);
+  }
+
+  goToMerchantOrders() {
+    this.router.navigate(['merchant-orders']);
+  }
+
+  goToMyProfile() {
+    this.router.navigate(['merchant-profile']);
   }
 
 }
